feat: moves wallpapers download to s3

+++ system-brew.sh
@@ -307,6 +307,7 @@ install_walls() {
     term_message cb "\nInstalling wallpapers..."
 
     target_dir="$HOME/Pictures"
+    walls_zip_filename="Walls.zip"
 
     if [ -d "${target_dir}/Walls/" ]; then
         local response
@@ -318,32 +319,68 @@ install_walls() {
                 break
                 ;;
             *)
-                echo
+                echo "Skipping wallpapers installation"
                 exit
                 ;;
             esac
         done
     fi
 
-    task_start "Downloading wallpapers"
-    curl -L "https://drive.google.com/file/d/1SLGW6dZ_6vEZp4BOfKXnBo9uKoM2r7wD/view?usp=drive_link" -o ${target_dir}/Walls.zip
-    task_done "Downloaded wallpapers"
-
-    task_start "Unzipping wallpapers"
-    unzip -d ${target_dir}/Walls ${target_dir}/Walls.zip
-    task_done "Unzipped wallpapers"
-
-    task_start "Removing .zip file"
-    rm ${target_dir}/Walls.zip
-    task_done "Removed .zip file"
-
-    if [ -d "${target_dir}/Walls/__MACOSX" ]; then
-        task_start "Removing __MACOSX folder"
-        rm -r ${target_dir}/Walls/__MACOSX
-        task_done "Removed __MACOSX folder"
+    task_start "Downloading wallpapers...\n"
+    if [ -f "${target_dir}/${walls_zip_filename}" ]; then
+        local response
+        while true; do
+            read -r -p "There is already a wallpapers zip file in ${target_dir} Do you want to continue? (y/n) " response
+            case "${response}" in
+            [yY][eE][sS] | [yY])
+                echo
+                break
+                ;;
+            *)
+                echo "Did not download the wallpapers since a .zip file with the same name already exists in ${target_dir}"
+                exit
+                ;;
+            esac
+        done
     fi
 
-    term_message gb "\nWallpapers installed."
+    if curl -L "https://s3.us-west-2.amazonaws.com/demo.yujinelson.com/${walls_zip_filename}" -o ${target_dir}/${walls_zip_filename}; then
+        task_done "Downloaded wallpapers"
+    else
+        task_fail "Could not download wallpapers"
+    fi
+
+    if [ -f "${target_dir}/${walls_zip_filename}" ]; then
+        task_start "Unzipping wallpapers..."
+        if unzip -q ${target_dir}/${walls_zip_filename} -d ${target_dir}/Walls; then
+            task_done "Unzipped wallpapers to ${target_dir}/Walls"
+        else
+            task_fail "Could not unzip wallpaper"
+        fi
+
+        task_start "Removing .zip file"
+        rm ${target_dir}/${walls_zip_filename}
+        task_done "Removed .zip file"
+
+        if [ -d "${target_dir}/Walls/Walls" ]; then
+            task_start "Cleaning up Walls/Walls folder..."
+            mv -i $target_dir/Walls/Walls/* $target_dir/Walls/
+
+            if rm -r ${target_dir}/Walls/Walls; then
+                task_done "Removed Walls/Walls folder"
+            else
+                task_fail "Could not automatically remove Walls/Walls folder"
+            fi
+        fi
+
+        if [ -d "${target_dir}/Walls/__MACOSX" ]; then
+            task_start "Removing __MACOSX folder..."
+            rm -r ${target_dir}/Walls/__MACOSX
+            task_done "Removed __MACOSX folder"
+        fi
+
+        term_message gb "\nWallpapers installed."
+    fi
 }
 
 install_lockscreen_image() {
